setup/requirements.txt
tornado == 6.1
tornado-sqlalchemy == 0.7.0
psycopg2-binary == 2.9.3